Updated AboutPageTop component & Fixed font-size & moved PageTopSection to shared
