+++ src/components/shared/PageTopSection.tsx
@@ -6,10 +6,10 @@ import React from "react";
 
 import { localeInURL } from "@/src/utils/localeInURL";
 
-import { GoBackButton } from "../shared/GoBackButton";
-import { IconHome } from "../shared/Icons/IcomHome";
-import LocaleSwitcher from "../shared/LocaleSwitcher";
-import { ThemeSwitcher } from "../shared/ThemeSwitcher";
+import { GoBackButton } from "./GoBackButton";
+import { IconHome } from "./Icons/IcomHome";
+import LocaleSwitcher from "./LocaleSwitcher";
+import { ThemeSwitcher } from "./ThemeSwitcher";
 
 interface PageTopSectionProps {
     breadcrumbsList: { title: string; url: string }[];
